requirements.txt
pyairtable==2.3.3
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.8.2